test_requirements.txt
# -e git+http://github.com/aldryn/aldryn-apphooks-config.git#egg=aldryn-apphooks-config
coverage==3.7.1
python-coveralls==2.4.2
django>=1.6,<1.7
django-admin-sortable
django-classy-tags
# -e git+http://github.com/yakky/django-cms/@feature/appspaced_apphooks#egg=django-cms
# Looking to remove this dependency!
django-hvad
django-mptt
# Looking to add this soon
# django-parler
django-reversion>=1.8.2,<1.9
django-sekizai
django-sortedm2m
djangocms-admin-style
djangocms-helper
djangocms_text_ckeditor
docopt
html5lib
flake8
selenium
six
South
-e git+https://github.com/aldryn/aldryn-search.git#egg=aldryn-search
